statistik hafiz

+++ app/Http/Controllers/HafizController.php
@@ -44,9 +44,93 @@ class HafizController extends Controller
         return response()->json($hafizes);
     }
 
+    public function detail($id = 0)
+    {
+        $hafiz = Hafiz::findOrFail($id);
+        $memorized_surah_count = 0;
+        $memorized_ayah_count = 0;
+        $global_score = 0;
+        $surah_ids = [];
+        $result = DB::select('
+        SELECT s.id as surah_id, a.number as ayah_number, md.score
+        FROM memorization_details md
+        JOIN (
+            SELECT md.ayah_id, MAX(m.created_at) AS latest
+            FROM memorization_details md
+            JOIN memorizations m ON md.memorization_id = m.id
+            WHERE m.hafiz_id = :hafiz_id
+            GROUP BY md.ayah_id
+        ) recent ON md.ayah_id = recent.ayah_id
+        AND m.created_at = recent.latest
+        JOIN memorizations m ON md.memorization_id = m.id
+        JOIN ayahs a ON md.ayah_id = a.id
+        JOIN surahs s ON s.id = a.surah_id
+        WHERE m.hafiz_id = :hafiz_id
+        ', [$hafiz->id]);
+        $global_score_count = count($result);
+
+        $details = [];
+        $total_score = 0;
+        foreach ($result as $row) {
+            $surah_ids[$row->surah_id] = $row->surah_id;
+            if (!isset($details[$row->surah_id])) {
+                $total_score = 0;
+                $details[$row->surah_id] = [
+                    'surah_id' => $row->surah_id,
+                    'memorized_ayah_count' => 0,
+                    'average_score' => 0,
+                    'details' => [],
+                ];
+            }
+
+            $total_score += $row->score;
+
+            $details[$row->surah_id]['details'][$row->ayah_number] = $row->score;
+            $details[$row->surah_id]['memorized_ayah_count'] += 1;
+
+            $global_score += $row->score;
+        }
+        ksort($details);
+
+        $surahs = DB::table('surahs')
+            ->whereIn('id', array_keys($surah_ids))
+            ->orderBy('id', 'asc')
+            ->get();
+
+        foreach ($surahs as $surah) {
+            $details[$surah->id]['ayah_count'] = $surah->total_ayahs;
+            $details[$surah->id]['surah_name'] = $surah->name;
+            $total_score = 0;
+            foreach ($details[$surah->id]['details'] as $ayah_id => $score) {
+                $total_score += $score;
+                $memorized_ayah_count++;
+            }
+            $details[$surah->id]['average_score'] = $total_score / count($details[$surah->id]['details']);
+            $memorized_surah_count++;
+        }
+
+        $hafiz->average_score = 0;
+        if ($global_score_count > 0) {
+            $hafiz->average_score = $global_score / $global_score_count;
+        }
+
+        $hafiz->memorized_surah_count = $memorized_surah_count;
+        $hafiz->memorized_ayah_count = $memorized_ayah_count;
+
+        return inertia('hafiz/Detail', [
+            'data' => $hafiz,
+            'details' => $details,
+        ]);
+    }
+
     public function editor($id = 0)
     {
         $hafiz = $id ? Hafiz::findOrFail($id) : new Hafiz(['active' => true]);
+
+        if ($id && $hafiz->user_id != Auth::user()->id) {
+            return response()->json(['error' => 'You are not authorized to edit this hafiz'], 403);
+        }
+
         $juzs = HafizMemorizedJuz::where('hafiz_id', $hafiz->id)->get()->toArray();
         $surahs = HafizMemorizedSurah::where('hafiz_id', $hafiz->id)->get()->toArray();
         $data = $hafiz->toArray();
@@ -60,8 +144,10 @@ class HafizController extends Controller
 
     public function save(Request $request)
     {
-        $userId = Auth::user()->id;
-        $hafiz = !$request->id ? new Hafiz(['user_id' => $userId]) : Hafiz::findOrFail($request->id);
+        $hafiz = !$request->id ? new Hafiz(['user_id' => Auth::user()->id]) : Hafiz::findOrFail($request->id);
+        if ($hafiz->user_id != Auth::user()->id) {
+            return response()->json(['error' => 'You are not authorized to modify this hafiz'], 403);
+        }
 
         $request->validate([
             'name' => 'required|max:255',
@@ -85,7 +171,7 @@ class HafizController extends Controller
                 'hafiz_id' => $hafiz->id,
                 'juz' => $num,
             ]);
-            $juz ->save();
+            $juz->save();
         }
 
         HafizMemorizedSurah::where('hafiz_id', $hafiz->id)->delete();
@@ -94,20 +180,30 @@ class HafizController extends Controller
                 'hafiz_id' => $hafiz->id,
                 'surah_id' => $surah_id,
             ]);
-            $surah ->save();
+            $surah->save();
         }
         DB::commit();
-
         return redirect(route('hafiz.index'))->with('success', "Hafiz {$hafiz->name} telah disimpan.");
     }
 
     public function delete($id)
     {
         $hafiz = Hafiz::findOrFail($id);
+        if ($hafiz->user_id != Auth::user()->id) {
+            return response()->json(['error' => 'You are not authorized to delete this hafiz'], 403);
+        }
         $hafiz->delete();
+        return redirect(route('hafiz.index'))->with('success', "Hafiz {$hafiz->name} telah dihapus.");
+    }
 
-        return response()->json([
-            'message' => "Hafiz {$hafiz->name} telah dihapus!"
-        ]);
+    public function clearScore($id)
+    {
+        $hafiz = Hafiz::findOrFail($id);
+        if ($hafiz->user_id != Auth::user()->id) {
+            return response()->json(['error' => 'You are not authorized to reset this hafiz'], 403);
+        }
+
+        DB::delete('delete from memorizations where hafiz_id = ?', [$hafiz->id]);
+        return redirect(route('hafiz.index'))->with('success', "Riwayat hafalan hafiz {$hafiz->name} telah dihapus.");
     }
 }
